(maint) Update to clj-parent 4.8.1
This commit updates to the latest version of clj-parent, just to stay current.
This also resolves some false positives found by the Snyk security scanner,
related to outdated dependencies in the old version of clj-parent (the version
of clj-parent we actually ship has been up-to-date for quite some time, the
version in this repo is just used for clojure tests).